+++ apps/web/src/app/admin-console/organizations/members/components/account-recovery/index.ts
@@ -1,2 +1,1 @@
-export * from "./account-recovery-dialog.component";
 export * from "./account-recovery-dialog-v2.component";
