+++ apps/web/src/app/admin-console/organizations/members/members.component.spec.ts
@@ -34,7 +34,7 @@ import { OrganizationWarningsService } from "@bitwarden/web-vault/app/billing/or
 
 import { OrganizationUserView } from "../core/views/organization-user.view";
 
-import { AccountRecoveryDialogResultType } from "./components/account-recovery/account-recovery-dialog.component";
+import { AccountRecoveryDialogResultType } from "./components/account-recovery";
 import { MemberDialogResult } from "./components/member-dialog";
 import { MembersComponent } from "./members.component";
 import {
